+++ pkg/contracts/uniswap/v2-core/contracts/libraries/uq112x112.sol/uq112x112.go
@@ -1,0 +1,203 @@
+// Code generated - DO NOT EDIT.
+// This file is a generated binding and any manual changes will be lost.
+
+package uq112x112
+
+import (
+	"errors"
+	"math/big"
+	"strings"
+
+	ethereum "github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/event"
+)
+
+// Reference imports to suppress errors if they are not otherwise used.
+var (
+	_ = errors.New
+	_ = big.NewInt
+	_ = strings.NewReader
+	_ = ethereum.NotFound
+	_ = bind.Bind
+	_ = common.Big1
+	_ = types.BloomLookup
+	_ = event.NewSubscription
+	_ = abi.ConvertType
+)
+
+// UQ112x112MetaData contains all meta data concerning the UQ112x112 contract.
+var UQ112x112MetaData = &bind.MetaData{
+	ABI: "[]",
+	Bin: "0x60556023600b82828239805160001a607314601657fe5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea265627a7a72315820c4e545b71eb450fec89d3fbcceea2f62f8971fce94db924228248b088e907ed864736f6c63430005100032",
+}
+
+// UQ112x112ABI is the input ABI used to generate the binding from.
+// Deprecated: Use UQ112x112MetaData.ABI instead.
+var UQ112x112ABI = UQ112x112MetaData.ABI
+
+// UQ112x112Bin is the compiled bytecode used for deploying new contracts.
+// Deprecated: Use UQ112x112MetaData.Bin instead.
+var UQ112x112Bin = UQ112x112MetaData.Bin
+
+// DeployUQ112x112 deploys a new Ethereum contract, binding an instance of UQ112x112 to it.
+func DeployUQ112x112(auth *bind.TransactOpts, backend bind.ContractBackend) (common.Address, *types.Transaction, *UQ112x112, error) {
+	parsed, err := UQ112x112MetaData.GetAbi()
+	if err != nil {
+		return common.Address{}, nil, nil, err
+	}
+	if parsed == nil {
+		return common.Address{}, nil, nil, errors.New("GetABI returned nil")
+	}
+
+	address, tx, contract, err := bind.DeployContract(auth, *parsed, common.FromHex(UQ112x112Bin), backend)
+	if err != nil {
+		return common.Address{}, nil, nil, err
+	}
+	return address, tx, &UQ112x112{UQ112x112Caller: UQ112x112Caller{contract: contract}, UQ112x112Transactor: UQ112x112Transactor{contract: contract}, UQ112x112Filterer: UQ112x112Filterer{contract: contract}}, nil
+}
+
+// UQ112x112 is an auto generated Go binding around an Ethereum contract.
+type UQ112x112 struct {
+	UQ112x112Caller     // Read-only binding to the contract
+	UQ112x112Transactor // Write-only binding to the contract
+	UQ112x112Filterer   // Log filterer for contract events
+}
+
+// UQ112x112Caller is an auto generated read-only Go binding around an Ethereum contract.
+type UQ112x112Caller struct {
+	contract *bind.BoundContract // Generic contract wrapper for the low level calls
+}
+
+// UQ112x112Transactor is an auto generated write-only Go binding around an Ethereum contract.
+type UQ112x112Transactor struct {
+	contract *bind.BoundContract // Generic contract wrapper for the low level calls
+}
+
+// UQ112x112Filterer is an auto generated log filtering Go binding around an Ethereum contract events.
+type UQ112x112Filterer struct {
+	contract *bind.BoundContract // Generic contract wrapper for the low level calls
+}
+
+// UQ112x112Session is an auto generated Go binding around an Ethereum contract,
+// with pre-set call and transact options.
+type UQ112x112Session struct {
+	Contract     *UQ112x112        // Generic contract binding to set the session for
+	CallOpts     bind.CallOpts     // Call options to use throughout this session
+	TransactOpts bind.TransactOpts // Transaction auth options to use throughout this session
+}
+
+// UQ112x112CallerSession is an auto generated read-only Go binding around an Ethereum contract,
+// with pre-set call options.
+type UQ112x112CallerSession struct {
+	Contract *UQ112x112Caller // Generic contract caller binding to set the session for
+	CallOpts bind.CallOpts    // Call options to use throughout this session
+}
+
+// UQ112x112TransactorSession is an auto generated write-only Go binding around an Ethereum contract,
+// with pre-set transact options.
+type UQ112x112TransactorSession struct {
+	Contract     *UQ112x112Transactor // Generic contract transactor binding to set the session for
+	TransactOpts bind.TransactOpts    // Transaction auth options to use throughout this session
+}
+
+// UQ112x112Raw is an auto generated low-level Go binding around an Ethereum contract.
+type UQ112x112Raw struct {
+	Contract *UQ112x112 // Generic contract binding to access the raw methods on
+}
+
+// UQ112x112CallerRaw is an auto generated low-level read-only Go binding around an Ethereum contract.
+type UQ112x112CallerRaw struct {
+	Contract *UQ112x112Caller // Generic read-only contract binding to access the raw methods on
+}
+
+// UQ112x112TransactorRaw is an auto generated low-level write-only Go binding around an Ethereum contract.
+type UQ112x112TransactorRaw struct {
+	Contract *UQ112x112Transactor // Generic write-only contract binding to access the raw methods on
+}
+
+// NewUQ112x112 creates a new instance of UQ112x112, bound to a specific deployed contract.
+func NewUQ112x112(address common.Address, backend bind.ContractBackend) (*UQ112x112, error) {
+	contract, err := bindUQ112x112(address, backend, backend, backend)
+	if err != nil {
+		return nil, err
+	}
+	return &UQ112x112{UQ112x112Caller: UQ112x112Caller{contract: contract}, UQ112x112Transactor: UQ112x112Transactor{contract: contract}, UQ112x112Filterer: UQ112x112Filterer{contract: contract}}, nil
+}
+
+// NewUQ112x112Caller creates a new read-only instance of UQ112x112, bound to a specific deployed contract.
+func NewUQ112x112Caller(address common.Address, caller bind.ContractCaller) (*UQ112x112Caller, error) {
+	contract, err := bindUQ112x112(address, caller, nil, nil)
+	if err != nil {
+		return nil, err
+	}
+	return &UQ112x112Caller{contract: contract}, nil
+}
+
+// NewUQ112x112Transactor creates a new write-only instance of UQ112x112, bound to a specific deployed contract.
+func NewUQ112x112Transactor(address common.Address, transactor bind.ContractTransactor) (*UQ112x112Transactor, error) {
+	contract, err := bindUQ112x112(address, nil, transactor, nil)
+	if err != nil {
+		return nil, err
+	}
+	return &UQ112x112Transactor{contract: contract}, nil
+}
+
+// NewUQ112x112Filterer creates a new log filterer instance of UQ112x112, bound to a specific deployed contract.
+func NewUQ112x112Filterer(address common.Address, filterer bind.ContractFilterer) (*UQ112x112Filterer, error) {
+	contract, err := bindUQ112x112(address, nil, nil, filterer)
+	if err != nil {
+		return nil, err
+	}
+	return &UQ112x112Filterer{contract: contract}, nil
+}
+
+// bindUQ112x112 binds a generic wrapper to an already deployed contract.
+func bindUQ112x112(address common.Address, caller bind.ContractCaller, transactor bind.ContractTransactor, filterer bind.ContractFilterer) (*bind.BoundContract, error) {
+	parsed, err := UQ112x112MetaData.GetAbi()
+	if err != nil {
+		return nil, err
+	}
+	return bind.NewBoundContract(address, *parsed, caller, transactor, filterer), nil
+}
+
+// Call invokes the (constant) contract method with params as input values and
+// sets the output to result. The result type might be a single field for simple
+// returns, a slice of interfaces for anonymous returns and a struct for named
+// returns.
+func (_UQ112x112 *UQ112x112Raw) Call(opts *bind.CallOpts, result *[]interface{}, method string, params ...interface{}) error {
+	return _UQ112x112.Contract.UQ112x112Caller.contract.Call(opts, result, method, params...)
+}
+
+// Transfer initiates a plain transaction to move funds to the contract, calling
+// its default method if one is available.
+func (_UQ112x112 *UQ112x112Raw) Transfer(opts *bind.TransactOpts) (*types.Transaction, error) {
+	return _UQ112x112.Contract.UQ112x112Transactor.contract.Transfer(opts)
+}
+
+// Transact invokes the (paid) contract method with params as input values.
+func (_UQ112x112 *UQ112x112Raw) Transact(opts *bind.TransactOpts, method string, params ...interface{}) (*types.Transaction, error) {
+	return _UQ112x112.Contract.UQ112x112Transactor.contract.Transact(opts, method, params...)
+}
+
+// Call invokes the (constant) contract method with params as input values and
+// sets the output to result. The result type might be a single field for simple
+// returns, a slice of interfaces for anonymous returns and a struct for named
+// returns.
+func (_UQ112x112 *UQ112x112CallerRaw) Call(opts *bind.CallOpts, result *[]interface{}, method string, params ...interface{}) error {
+	return _UQ112x112.Contract.contract.Call(opts, result, method, params...)
+}
+
+// Transfer initiates a plain transaction to move funds to the contract, calling
+// its default method if one is available.
+func (_UQ112x112 *UQ112x112TransactorRaw) Transfer(opts *bind.TransactOpts) (*types.Transaction, error) {
+	return _UQ112x112.Contract.contract.Transfer(opts)
+}
+
+// Transact invokes the (paid) contract method with params as input values.
+func (_UQ112x112 *UQ112x112TransactorRaw) Transact(opts *bind.TransactOpts, method string, params ...interface{}) (*types.Transaction, error) {
+	return _UQ112x112.Contract.contract.Transact(opts, method, params...)
+}