+++ zetaclient/chains/evm/signer/signer_admin.go
@@ -8,7 +8,7 @@ import (
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/zeta-chain/protocol-contracts/v1/pkg/contracts/evm/erc20custody.sol"
+	"github.com/zeta-chain/protocol-contracts/pkg/erc20custody.sol"
 
 	"github.com/zeta-chain/node/pkg/constant"
 )